+++ client/src/components/Header.js
@@ -1,12 +1,12 @@
-import React, { useState } from "react";
+import React from "react";
 
-import { Box, Typography, Tabs, Tab } from "@material-ui/core";
+import { Box } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 
 import logo from "img/logo.png";
 
-import Notifications from "components/Notifications";
 import ProfileMenu from "components/ProfileMenu";
+import HeaderTabs from "components/HeaderTabs";
 
 const useStyles = makeStyles((theme) => ({
   headerContainer: {
@@ -17,36 +17,6 @@ const useStyles = makeStyles((theme) => ({
     justifyContent: "space-evenly",
     boxSizing: "border-box",
   },
-  dashboardHeader: {
-    height: "auto",
-    width: "100%",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "space-evenly",
-    boxSizing: "border-box",
-  },
-  linkContainer: {
-    marginLeft: theme.spacing(6),
-    marginRight: theme.spacing(6),
-    display: "flex",
-    alignItems: "center",
-  },
-  link: {
-    margin: theme.spacing(3),
-    textDecoration: "none",
-    color: "black",
-    position: "relative",
-    padding: "5px 18px",
-  },
-  notificationIcon: {
-    top: "0",
-    right: "0",
-    height: "10px",
-    width: "10px",
-    backgroundColor: "#35d554",
-    borderRadius: "100%",
-    position: "absolute",
-  },
   logo: {
     height: "1.8rem",
     padding: theme.spacing(1),
@@ -56,32 +26,12 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 const Header = (props) => {
-  const { selectedPage, setSelectedPage } = props;
   const classes = useStyles();
-
-  const handleTabChange = (event, newValue) => {
-    setSelectedPage(newValue);
-  };
 
   return (
     <Box className={classes.headerContainer}>
       <img className={classes.logo} src={logo} alt="logo" />
-      <Box className={classes.linkContainer}>
-        <Tabs
-          value={selectedPage}
-          onChange={handleTabChange}
-          aria-label="page select"
-          TabIndicatorProps={{ style: { backgroundColor: "white" } }}
-        >
-          <Tab label="Shopping Lists" />
-          <Tab label="Friends" />
-        </Tabs>
-        <Box className={classes.link}>
-          <Typography>Notifications</Typography>
-          <Box className={classes.notificationIcon} />
-          <Notifications />
-        </Box>
-      </Box>
+      <HeaderTabs {...props} />
       <ProfileMenu />
     </Box>
   );
